+++ src/components/posts.js
@@ -2,7 +2,7 @@ import Post from './post';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import { db } from './../firebase';
-import { ref, onValue } from 'firebase/database';
+import { ref, onValue, get } from 'firebase/database';
 import { useState, useEffect } from 'react';
 
 
@@ -12,7 +12,7 @@ function Posts() {
 
     useEffect(() => {
         const postRef = ref(db, 'posts');
-        return onValue(postRef, snapshots => {
+        get(postRef).then(snapshots => {
            snapshots.forEach(snapshot => {
                const data = snapshot.val();
                 setposts(prevPosts => [...prevPosts, data]);
@@ -28,7 +28,7 @@ function Posts() {
                     spacing={1}
                 >
                     {
-                        posts.map((post, index) => 
+                        posts.reverse().map((post, index) =>
                             <Grid item key={`post${index}`} xs={12} md={6}>
                                 <Post post={post} />
                             </Grid>
